MAINT: delete duplictated functions
