+++ tests/test_dbs.py
@@ -27,3 +27,19 @@ def test_blob_array(data):
     # handles array or bytes as input
     inflated = eti_storage.blob_to_array(data)
     assert numpy.array_equal(inflated, numpy.array([0, 3], dtype=numpy.uint8))
+
+
+@pytest.mark.parametrize(
+    "data",
+    [
+        numpy.array([1, 2, 3], dtype=numpy.int32),
+        numpy.array([[1, 2, 3], [4, 5, 6]], dtype=numpy.int32),
+        numpy.array([[1, 2, 3], [4, 5, 6]], dtype=numpy.uint8),
+    ],
+)
+def test_array_blob_interconvert(data):
+    blob = eti_storage.array_to_blob(data)
+    assert isinstance(blob, bytes)
+    back = eti_storage.blob_to_array(blob)
+    assert numpy.array_equal(data, back)
+    assert data.dtype == back.dtype
